+++ examples/OpenGLWindow/CMakeLists.txt
@@ -1,6 +1,7 @@
 
 INCLUDE_DIRECTORIES(
 	..
+	../ThirdPartyLibs
 	../../src
 )
