+++ examples/OpenGLWindow/GLInstancingRenderer.cpp
@@ -1406,7 +1406,7 @@ void GLInstancingRenderer::updateCamera(int upAxis)
 
 
 //#define STB_IMAGE_WRITE_IMPLEMENTATION
-#include "stb_image_write.h"
+#include "stb_image/stb_image_write.h"
 void writeTextureToPng(int textureWidth, int textureHeight, const char* fileName, int numComponents)
 {
