+++ examples/OpenGLWindow/fontstash.cpp
@@ -37,7 +37,7 @@
 #define STB_TRUETYPE_IMPLEMENTATION
 #define STBTT_malloc(x,u)    malloc(x)
 #define STBTT_free(x,u)      free(x)
-#include "stb_truetype.h"
+#include "stb_image/stb_truetype.h"
 
 #define HASH_LUT_SIZE 256
