+++ examples/OpenGLWindow/opengl_fontstashcallbacks.cpp
@@ -10,7 +10,7 @@
 #include <stdlib.h>
 #include <string.h>
 #define STB_IMAGE_WRITE_IMPLEMENTATION
-#include "stb_image_write.h"
+#include "stb_image/stb_image_write.h"
 
 
 static unsigned int s_indexData[INDEX_COUNT];
