+++ examples/ThirdPartyLibs/stb_image/stb_image_write.h
